Refactor SaveTagAsync method parameters and logic
Updated the `SaveTagAsync` method in the `SqliteConnectionFactory` class to accept a `string` for the tag text and a `uint` for the trainer's ID instead of a `Tags` object. A new `Tags` object is now created within the method using the provided parameters to facilitate saving the tag to the database.