Remove Android and Shared folder references
Updated `PokemonBattleJournal.UI.Tests.csproj` to eliminate all references to the `Android` and `Shared` folders. This includes the removal of related compile, embedded resource, Maui CSS, Maui XAML, and none items, as well as the deletion of the corresponding `ItemGroup`. This change reflects a restructuring of the project to streamline the build process.